fixing bug in profiling in which int's are handled improperly (handled as floating point numbers w/o proper casts), resulting in erroneous data and/or loss of precision.
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@66867 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acstime/ws/src/alma/acs/time/Profiler.java
@@ -95,7 +95,7 @@ public class Profiler extends alma.acs.util.StopWatch
 	    String out = "";
 	    
 	    out = out + "#ACS PROFILER# msg=" + msg;
-	    out = out + ", avg=" + m_totalTime /  m_totalNumStarts; 
+	    out = out + ", avg=" + (double)((double)m_totalTime /  (double)m_totalNumStarts); 
 	    out = out + ", runs=" + m_totalNumStarts; 
 	    out = out + ", mindur=" + m_minDuration;
 	    out = out + ", maxdur=" + m_maxDuration;
